Added OptValue - a value with optional contents.
I'm not thrilled at the level of code duplication here, but I also don't want
to sacrifice performance in this largely stable codebase.

I'm also not thrilled about the creation of a Some every time the value is
updated (since other reactors can do zero allocation event dispatch), but it's
no worse than a Value[Option[T]] and this API is nicer.